Reworked Header and Nav - but need to rework wrap
I got this far, having reworked the header to include H1 for site
title, and allowing the Nav to be it’s own thing. But as I work on the
desktop layout - I’m starting to think I should have wrapped everything
in a single wrap div - rather than add .wrap to multiple classes. So
that’s what I’m going to do next.